refactor: change folder structure to optimise style files refactor: change components names to be semantically right
