+++ src/components/Layouts/Layout.js
@@ -2,11 +2,10 @@ import React from "react"
 import PropTypes from "prop-types"
 import { Container, Row, Col } from 'react-bootstrap'
 
-import Header from "./header"
-import footerStyles from '../styles/layout.module.less'
-import mainStyles from '../styles/main.module.less'
-import "../styles/layout.less"
-import "../styles/style.less"
+import Header from "../Header/Header"
+import footerStyles from './layout.module.less'
+import mainStyles from '../../mainStyles/main.module.less'
+import "./layout.less"
 
 const Layout = ({ children }) => {
   return (
