--- src/styles/style.less
@@ -1,3 +1,0 @@
-* {
-	// color: red;
-}
